Remove oliver and *.cs.colostate.edu from the dotfiles.
The oliver VPS at DigitalOcean has been deleted.

Since graduating from CSU, access will soon be revoked.
Until it is revoked, keep the SSH key and the SSH config.
Oliver's SSH key has been removed.

The NSS module is not necessary without the CS machines.
This also opened up the oversight that the SSL module was
no longer needed. The certificate has not been used since
moving over to a web-based solution and thus, the entire
module can be deleted.

The colostate module has been deleted as it is no longer
useful and has not been operational in quite some time.

The color for hastings has been changed to green.